+++ CHANGES.rst
@@ -3,7 +3,7 @@ Version history
 
 **UNRELEASED**
 
-Add- Fixed Enum columns code generation
+- Support optional foreign key relationships
 - Fixed SQLModel code generation
 - Dropped support for Python 3.8
 - Fixed two rendering issues in ``ENUM`` columns when a non-default schema is used: an